Changes to be committed: 	modified:   resonator_tools/circuit.py
added method 'convert_fitresults_to_lambda_half_hanger' to 'notch_port' class. The conversion which is being
performed by the new method enables the user to use the same fitting procedure for a half-wavelength resonator
as for the quarter wavelength resonator. Afterwards, the fit parameters are converted into the effective fit
parameters of a half-wavelength resonator.